fix: address review feedback
- Fix Y2038 timestamp truncation: sender_key_devices.updated_at uses
  i64/BigInt instead of i32/Integer (pre-existing i32 in device_registry
  and base_keys left for separate fix)
- Remove leftover comments referencing old API methods
- Log warning on DB error in resolve_skdm_targets instead of silently
  falling through to full SKDM distribution

+++ storages/sqlite-storage/src/schema.rs
@@ -106,7 +106,7 @@ diesel::table! {
         device_jid -> Text,
         has_key -> Integer,
         device_id -> Integer,
-        updated_at -> Integer,
+        updated_at -> BigInt,
     }
 }
